Move reasoningEffort.ts from agent/runtime to modelHandlers/support
This file maps user-facing reasoning levels to provider API effort
values. Its primary consumers are the OpenAI model handlers
(modelHandlerOpenAI, modelHandlerOpenAIResponse) and the settings
controller, not execution-orchestration code. Placing it in
modelHandlers/support aligns it with the other model-handler support
utilities (sdkErrorMetadata, ProxyConfigResolver, etc.) and removes
a model-handler-specific concern from the agent/runtime flat namespace.
